Store gauge outputs as complex64
Save gauge configurations and Polyakov correlator outputs with complex64 precision to reduce disk usage. Adjust SU(3) validation tolerances and tests for single-precision persisted links.

Validation: UV_CACHE_DIR=/tmp/uv-cache uv run pytest

+++ scripts/measure_polyakov_correlators.py
@@ -202,7 +202,7 @@ def measure_correlator_ensemble(
         filenames.append(path.name)
 
     return (
-        np.asarray(correlators, dtype=np.complex128),
+        np.asarray(correlators, dtype=np.complex64),
         np.asarray(sweeps, dtype=np.int64),
         np.asarray(chains, dtype=np.int64),
         np.asarray(starts, dtype=np.str_),
